Added UI scaling to elements in the game overlay and pause support
UI now scales with resolution while playing.
Press start to pause the game with a controller or "P" to pause the game
with a keyboard.
PS3/4 Controller support is currently not mapping to the right key so
right now the right trigger is what Unity currently thinks is the start
button.